Move printVersions function to version.js
